refactor: Move component directories into Calendar page directory
