--- packages/mobile/src/components/atoms/AddScheduleLink/style.module.scss
@@ -1,3 +1,0 @@
-.link {
-    line-height: 0;
-}
