--- packages/mobile/src/components/atoms/DisplayDate/style.module.scss
@@ -1,6 +1,0 @@
-.year-month {
-    margin: 0 24px;
-    font-size: 18px;
-    font-weight: 500;
-    color: #202020;
-}
