+++ packages/mobile/src/page/Calendar/CollegeCard/style.module.scss
