+++ packages/mobile/src/page/Calendar/PersonalCard/style.module.scss
